Split lark inbound message handling tests
